Consistent naming and reasonable initial size
